refactor[openbeats]: update imports for new espnet2/beats/ layout
Follow the espnet2/beats/ relocation on PR #6436 — point the
pretraining model, BeatsTask, dump_audio_tokens.py, and unit tests
at the new module paths.
