+++ espnet2/beats/espnet_model.py
@@ -10,11 +10,11 @@ import torch.nn.functional as F
 from typeguard import typechecked
 
 from espnet2.asr.encoder.abs_encoder import AbsEncoder
-from espnet2.legacy.nets.pytorch_backend.nets_utils import make_pad_mask
-from espnet2.speechlm.tokenizer.beats_utils import (
+from espnet2.beats.utils import (
     beats_frontend,
     forward_padding_mask_conv,
 )
+from espnet2.legacy.nets.pytorch_backend.nets_utils import make_pad_mask
 from espnet2.torch_utils.device_funcs import force_gatherable
 from espnet2.train.abs_espnet_model import AbsESPnetModel
