+++ test/espnet2/beats/test_espnet_model.py
@@ -1,9 +1,9 @@
 import torch
 from packaging.version import parse as V
 
-from espnet2.asr.encoder.beats_encoder import BeatsEncoder, BeatsPretrainingPredictor
+from espnet2.beats.encoder import BeatsEncoder, BeatsPretrainingPredictor
 from espnet2.beats.espnet_model import BeatsPretrainModel, BeatsTokenizerPretrainModel
-from espnet2.speechlm.tokenizer.beats_tokenizer import (
+from espnet2.beats.tokenizer import (
     BeatsTokenizer,
     BeatsTokenizerPretrainingPredictor,
 )
